Medical-Image-Detection-App/requirements.txt
streamlit==1.30.0
google-generativeai
# Any other dependencies your app requires